Fix bug in seat position shifting on booking
