+++ src/components/BookNow.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-const BookNow = ({ price, selected }) => {
+const BookNow = ({ price, seatNo }) => {
   const styles = {
     background: "#F20C0C",
     fontWeight:"700",
@@ -14,8 +14,8 @@ const BookNow = ({ price, selected }) => {
   return (
     <div className="d-flex p-3 mt-5 justify-content-between" style={styles}>
       <div className="d-flex flex-column">
-        <h1 className="text-white">Total: ${price}</h1>
-        <p className="text-white">SEATS SELECTED: {selected}</p>
+        <h1 className="text-white">Total: Kshs. {seatNo? price: 0}</h1>
+        <p className="text-white">SEAT SELECTED: {seatNo}</p>
       </div>
       <div className="align-self-center">
         <button style={btnStyle} className="btn btn-link text-white">Book Now</button>
